Separated dual token test
